third_party/blink/web_tests/wpt_internal/atomic-move-auto/css-transition-to-disconnected-document.html

<!DOCTYPE html>
<title>Node.moveBefore should act like insertBefore when moving to a disconnected document</title>
<meta name="variant" content="?action=append">
<meta name="variant" content="?action=prepend">
<meta name="variant" content="?action=before">
<meta name="variant" content="?action=after">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helper.js"></script>

<body>
  <section id="old-parent">
    <div id="item"></div>
  </section>
  <section id="new-parent">
  </section>
  <style>
    #item {
      width: 100px;
      height: 100px;
      background: green;
      transition: left 10s;
      position: absolute;
      left: 0;
    }

    section {
      position: relative;
    }

    body {
      margin-left: 0;
    }
  </style>

  <script>
    promise_test(async t => {
      const item = document.querySelector("#item");
      assert_equals(item.getBoundingClientRect().x, 0);
      item.style.left = "400px";
      await new Promise(resolve => item.addEventListener("transitionstart", resolve));
      const doc = document.implementation.createHTMLDocument();
      reparent(doc.body, item);
      await new Promise(resolve => requestAnimationFrame(() => resolve()));
      assert_equals(item.getBoundingClientRect().x, 0);
    }, "Moving an element with a transition to a disconnected document should reset the transitionm state");
  </script>
</body>
